move status into repo minor notification fixes minor doc fixes minor style reformat
