+++ github4s/js/src/test/scala/github4s/integration/IntegrationSpec.scala
@@ -32,7 +32,6 @@ class IntegrationSpec
     with GHIssuesSpec[SimpleHttpResponse]
     with GHPullRequestsSpec[SimpleHttpResponse]
     with GHReposSpec[SimpleHttpResponse]
-    with GHStatusesSpec[SimpleHttpResponse]
     with GHUsersSpec[SimpleHttpResponse]
     with GHActivitiesSpec[SimpleHttpResponse]
     with ImplicitsJS
