+++ github4s/jvm/src/test/scala/github4s/integration/IntegrationSpec.scala
@@ -32,7 +32,6 @@ class IntegrationSpec
     with GHIssuesSpec[HttpResponse[String]]
     with GHPullRequestsSpec[HttpResponse[String]]
     with GHReposSpec[HttpResponse[String]]
-    with GHStatusesSpec[HttpResponse[String]]
     with GHUsersSpec[HttpResponse[String]]
     with GHActivitiesSpec[HttpResponse[String]]
     with ImplicitsJVM
