+++ github4s/jvm/src/test/scala/github4s/unit/ApiSpec.scala
@@ -40,7 +40,6 @@ class ApiSpec
   val gists        = new Gists[HttpResponse[String], Id]
   val gitData      = new GitData[HttpResponse[String], Id]
   val pullRequests = new PullRequests[HttpResponse[String], Id]
-  val statuses     = new Statuses[HttpResponse[String], Id]
   val issues       = new Issues[HttpResponse[String], Id]
   val activities   = new Activities[HttpResponse[String], Id]
 
@@ -316,6 +315,102 @@ class ApiSpec
         validTagTitle,
         validTagTitle,
         validNote)
+    response should be('left)
+  }
+
+  "Repos >> GetStatus" should "return the expected combined status when a valid ref is provided" in {
+    val response =
+      repos.get(accessToken, headerUserAgent, validRepoOwner, validRepoName, validRefSingle)
+    response should be('right)
+
+    response.toOption map { r ⇒
+      r.statusCode shouldBe okStatusCode
+    }
+  }
+
+  it should "return an error if no tokens are provided" in {
+    val response =
+      repos.get(None, headerUserAgent, validRepoOwner, validRepoName, validRefSingle)
+    response should be('left)
+  }
+
+  it should "return an error if an invalid ref is passed" in {
+    val response =
+      repos.get(accessToken, headerUserAgent, validRepoOwner, validRepoName, invalidRef)
+    response should be('left)
+  }
+
+  "Repos >> ListStatus" should "return the expected statuses when a valid ref is provided" in {
+    val response =
+      repos.listStatus(accessToken, headerUserAgent, validRepoOwner, validRepoName, validRefSingle)
+    response should be('right)
+
+    response.toOption map { r ⇒
+      r.result.nonEmpty shouldBe true
+      r.statusCode shouldBe okStatusCode
+    }
+  }
+
+  it should "return an error if no tokens are provided" in {
+    val response =
+      repos.listStatus(None, headerUserAgent, validRepoOwner, validRepoName, validRefSingle)
+    response should be('left)
+  }
+
+  it should "return an empty list when an invalid ref is passed" in {
+    val response =
+      repos.listStatus(accessToken, headerUserAgent, validRepoOwner, validRepoName, invalidRef)
+    response should be('right)
+
+    response.toOption map { r ⇒
+      r.result.isEmpty shouldBe true
+      r.statusCode shouldBe okStatusCode
+    }
+  }
+
+  "Repos >> CreateStatus" should "return the created status if a valid sha is provided" in {
+    val response = repos.createStatus(
+      accessToken,
+      headerUserAgent,
+      validRepoOwner,
+      validRepoName,
+      validCommitSha,
+      validStatusState,
+      None,
+      None,
+      None)
+    response should be('right)
+
+    response.toOption map { r =>
+      r.statusCode shouldBe createdStatusCode
+    }
+  }
+
+  it should "return an error if no tokens are provided" in {
+    val response = repos.createStatus(
+      None,
+      headerUserAgent,
+      validRepoOwner,
+      validRepoName,
+      validCommitSha,
+      validStatusState,
+      None,
+      None,
+      None)
+    response should be('left)
+  }
+
+  it should "return an error when an invalid sha is passed" in {
+    val response = repos.createStatus(
+      accessToken,
+      headerUserAgent,
+      validRepoOwner,
+      validRepoName,
+      invalidCommitSha,
+      validStatusState,
+      None,
+      None,
+      None)
     response should be('left)
   }
 
@@ -660,102 +755,6 @@ class ApiSpec
       invalidNewPullRequestIssue,
       invalidHead,
       invalidBase)
-    response should be('left)
-  }
-
-  "Statuses >> Get" should "return the expected combined status when a valid ref is provided" in {
-    val response =
-      statuses.get(accessToken, headerUserAgent, validRepoOwner, validRepoName, validRefSingle)
-    response should be('right)
-
-    response.toOption map { r ⇒
-      r.statusCode shouldBe okStatusCode
-    }
-  }
-
-  it should "return an error if no tokens are provided" in {
-    val response =
-      statuses.get(None, headerUserAgent, validRepoOwner, validRepoName, validRefSingle)
-    response should be('left)
-  }
-
-  it should "return an error if an invalid ref is passed" in {
-    val response =
-      statuses.get(accessToken, headerUserAgent, validRepoOwner, validRepoName, invalidRef)
-    response should be('left)
-  }
-
-  "Statuses >> List" should "return the expected statuses when a valid ref is provided" in {
-    val response =
-      statuses.list(accessToken, headerUserAgent, validRepoOwner, validRepoName, validRefSingle)
-    response should be('right)
-
-    response.toOption map { r ⇒
-      r.result.nonEmpty shouldBe true
-      r.statusCode shouldBe okStatusCode
-    }
-  }
-
-  it should "return an error if no tokens are provided" in {
-    val response =
-      statuses.list(None, headerUserAgent, validRepoOwner, validRepoName, validRefSingle)
-    response should be('left)
-  }
-
-  it should "return an empty list when an invalid ref is passed" in {
-    val response =
-      statuses.list(accessToken, headerUserAgent, validRepoOwner, validRepoName, invalidRef)
-    response should be('right)
-
-    response.toOption map { r ⇒
-      r.result.isEmpty shouldBe true
-      r.statusCode shouldBe okStatusCode
-    }
-  }
-
-  "Statuses >> Create" should "return the created status if a valid sha is provided" in {
-    val response = statuses.create(
-      accessToken,
-      headerUserAgent,
-      validRepoOwner,
-      validRepoName,
-      validCommitSha,
-      validStatusState,
-      None,
-      None,
-      None)
-    response should be('right)
-
-    response.toOption map { r =>
-      r.statusCode shouldBe createdStatusCode
-    }
-  }
-
-  it should "return an error if no tokens are provided" in {
-    val response = statuses.create(
-      None,
-      headerUserAgent,
-      validRepoOwner,
-      validRepoName,
-      validCommitSha,
-      validStatusState,
-      None,
-      None,
-      None)
-    response should be('left)
-  }
-
-  it should "return an error when an invalid sha is passed" in {
-    val response = statuses.create(
-      accessToken,
-      headerUserAgent,
-      validRepoOwner,
-      validRepoName,
-      invalidCommitSha,
-      validStatusState,
-      None,
-      None,
-      None)
     response should be('left)
   }
