+++ github4s/shared/src/main/scala/github4s/Github.scala
@@ -39,7 +39,6 @@ class Github(accessToken: Option[String] = None) {
   lazy val activities   = new GHActivities(accessToken)
   lazy val gitData      = new GHGitData(accessToken)
   lazy val pullRequests = new GHPullRequests(accessToken)
-  lazy val statuses     = new GHStatuses(accessToken)
 
 }
