+++ github4s/shared/src/main/scala/github4s/free/domain/Repository.scala
@@ -101,6 +101,45 @@ case class NewReleaseRequest(
     draft: Option[Boolean],
     prerelease: Option[Boolean])
 
+case class Status(
+    id: Int,
+    url: String,
+    state: String,
+    target_url: Option[String],
+    description: Option[String],
+    context: Option[String],
+    creator: Option[User],
+    created_at: String,
+    updated_at: String
+)
+
+case class NewStatusRequest(
+    state: String,
+    target_url: Option[String],
+    description: Option[String],
+    context: Option[String]
+)
+
+case class StatusRepository(
+    id: Int,
+    name: String,
+    full_name: String,
+    owner: User,
+    `private`: Boolean,
+    description: Option[String],
+    fork: Boolean,
+    urls: Map[String, String]
+)
+
+case class CombinedStatus(
+    url: String,
+    state: String,
+    commit_url: String,
+    sha: String,
+    total_count: Int,
+    statuses: List[Status],
+    repository: StatusRepository
+)
 object RepoUrlKeys {
 
   val forks_url         = "forks_url"
