+++ github4s/shared/src/test/scala/github4s/utils/BaseSpec.scala
@@ -119,7 +119,6 @@ trait BaseSpec extends FlatSpec with Matchers with TestData with IdInstances wit
   class GitDataOpsTest     extends GitDataOps[GitHub4s]
   class PullRequestOpsTest extends PullRequestOps[GitHub4s]
   class RepositoryOpsTest  extends RepositoryOps[GitHub4s]
-  class StatusOpsTest      extends StatusOps[GitHub4s]
   class IssueOpsTest       extends IssueOps[GitHub4s]
   class ActivityOpsTest    extends ActivityOps[GitHub4s]
